requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.2.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.2
multidict==6.0.4
pexpect==4.8.0
ptyprocess==0.7.0
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
six==1.16.0
Werkzeug==2.2.3
yarl==1.8.2
